+++ CHANGELOG.md
@@ -9,8 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added the ability to configure the underlying transport in Go. #1003
+
 ### Changed
 
+- Fixed an issue where multiple api clients could run into racing conditions in Go.
+- Fixed a bug where empty additional data in Go would lead to invalid JSON payloads during serialization.
+
+## [0.0.15] - 2021-12-17
+
+### Changed
+
+- Fixes name collisions in dotnet by renaming "HttpMethod" enum to "Method" in dotnet abstractions
+- Add support for PHP Generation.
 - Migrated generator to dotnet 6 #815
 - Fixes a bug where json deserialization would fail in go
 - Fixes a bug where query parameters would not be added to the request in go
@@ -21,6 +32,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrades go libraries to go17.
 - Fixes a bug in Go where reserved keywords for properties would be wrongly replaced.
 - Fixes a bug in Go where setters would be missing nil checks.
+- Fixes a bug where OData select query parameter would not be normalized
+- Fixes a bug in Go where empty collections would not be serialized.
+- Fixes a bug where generation would fail because of empty usings.
+- Fixes a bug where Java and Go escaped model properties would not serialize properly.
+- Fixes a bug where null values would not be added to additionalData if there was no matching property in dotnet. 
+- Fixes a bug where deserialzation of enums would throw an ArgumentExcpetion if the member didn't exist in dotnet.
 
 ## [0.0.14] - 2021-11-08